Dockerization (#17)
This PR adds support for deploying the webapp to a server using Docker and Docker-Compose. Along with the addition of the `Dockerfile` and `docker-compose.yml`, the code has been moved under the `app` directory, and some changes have been made in the templates in order to handle URLs correctly when running behind a reverse proxy mounted on a subdirectory of a domain (i.e. to make it run correctly when deployed to vanga.sista.arizona.edu/scivocab).

There are other minor refactorings as well to make the code more readable.